move asn1.rb contents to several files.
